+++ src/libmw/include/mw/node/State.h
@@ -24,9 +24,9 @@ struct State
 
         BitSet leafset = deserializer.Read<BitSet>();
 
-        const uint64_t num_utxos = deserializer.Read<uint64_t>();
+        const uint32_t num_utxos = deserializer.Read<uint32_t>();
         std::vector<UTXO::CPtr> utxos(num_utxos);
-        for (uint64_t i = 0; i < num_utxos; i++) {
+        for (uint32_t i = 0; i < num_utxos; i++) {
             utxos[i] = std::make_shared<UTXO>(deserializer.Read<UTXO>());
         }
